% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{items}
\alias{items}
\title{Scientific Publications and its Topics}
\format{
An object of class \code{"tbl_df"}
}
\source{
\url{https://github.com/dh-thesis/maxmodelr}
}
\usage{
data(items)
}
\description{
A data.frame (tibble) with publication identifiers and related topics.
}
\keyword{datasets}
