% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model}
\alias{model}
\title{Topic Model of Scientific Publications}
\format{
An object of class \code{"BTM"}
}
\source{
\url{https://github.com/dh-thesis/maxmodelr}
}
\usage{
data(model)
}
\description{
A biterm topic model of english publications from the Max Planck Society.
}
\keyword{datasets}
