% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar.R
\name{search_dist_items}
\alias{search_dist_items}
\title{Search for Items Most Likely Associated with Topic of Query Ranked by Distance}
\usage{
search_dist_items(query = "information retrieval", compare = 100, n = 10)
}
\arguments{
\item{query}{Query for topic inference.}

\item{compare}{Number of items to compare.}

\item{n}{Number of items to return.}
}
\description{
Search for Items Most Likely Associated with Topic of Query Ranked by Distance
}
