% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar.R
\name{search_topic_items}
\alias{search_topic_items}
\title{Search for Items Most Likely Associated with Topic of Query}
\usage{
search_topic_items(query = "information retrieval", n = 10)
}
\arguments{
\item{query}{Query for topic inference.}

\item{n}{Number of items to return.}
}
\description{
Search for Items Most Likely Associated with Topic of Query
}
