% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar.R
\name{search_topic_items_dist}
\alias{search_topic_items_dist}
\title{Search for Items by Topic Similarity of Query and Items Sharing the Most Likely Topic}
\usage{
search_topic_items_dist(query = "information retrieval", n = 10, limit = TRUE)
}
\arguments{
\item{query}{Query for topic inference.}

\item{n}{Number of items to return.}

\item{limit}{If vector size should be limited}
}
\description{
Search for Items by Topic Similarity of Query and Items Sharing the Most Likely Topic
}
