% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theta}
\alias{theta}
\title{Topic Distributions of Scientific Publications}
\format{
An object of class \code{"matrix"}
}
\source{
\url{https://github.com/dh-thesis/maxmodelr}
}
\usage{
data(theta)
}
\description{
A matrix with publications as rows and topic distributions as columns.
}
\keyword{datasets}
